+++ Packages/com.nickmaltbie.openkcc.netcode/Tests/Runtime/Character/NetworkKCCTests.cs
@@ -106,9 +106,10 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
             yield return base.UnitySetUp();
 
             // Spawn a floor below the players.
-            floor = GameObject.CreatePrimitive(PrimitiveType.Plane);
+            floor = GameObject.CreatePrimitive(PrimitiveType.Cube);
             floor.name = "floor";
-            floor.transform.localScale = new Vector3(10, 1, 10);
+            floor.transform.localScale = new Vector3(100, 1, 100);
+            floor.transform.position -= Vector3.up * 0.5f;
         }
 
         public override void SetupClient(NetworkKCC e, int objectIdx, int clientIdx)
@@ -137,7 +138,6 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
         [UnityTest]
         public IEnumerator Validate_NetworkKCC_MovingGround()
         {
-            floor.transform.localScale = Vector3.one;
             yield return SetupPlayersInIdleState();
 
             // Attach a moving ground conveyer to the ground
@@ -176,13 +176,13 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
         public IEnumerator Validate_NetworkKCC_Sliding()
         {
             SetupPlayersInIdleState();
+            floor.transform.position = Vector3.down * 3;
+            yield return new WaitForFixedUpdate();
+            floor.transform.rotation = Quaternion.Euler(70.0f, 0, 0);
 
-            floor.transform.rotation = Quaternion.Euler(61, 0, 0);
-            floor.transform.position += Vector3.back * 10 + Vector3.down * 5;
+            ForEachOwner((player, i) => player.TeleportPlayer(Vector3.zero));
 
-            // Teleport players back to their original positions
-            ForEachOwner((player, i) => player.TeleportPlayer(Vector3.right * i * 2 + Vector3.up * 0.0025f));
-
+            // Wait until players are in sliding satte
             yield return TestUtils.WaitUntil(() => ForAllPlayers(player => typeof(SlidingState) == player.CurrentState));
         }
 
@@ -204,7 +204,6 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
             ForEachOwner((player, i) =>
             {
                 player.TeleportPlayer(Vector3.right * i * 2 + Vector3.up * 0.0025f);
-                player.SetStateQuiet(typeof(IdleState));
             });
             SetupInputs();
             yield return TestUtils.WaitUntil(() => ForAllPlayers(player => typeof(IdleState) == player.CurrentState));
@@ -243,8 +242,16 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
 
         public override void SetupPrefab(GameObject go)
         {
+            NetworkKCC kcc = go.GetComponent<NetworkKCC>();
+            kcc.config.maxPushSpeed = 100.0f;
+
             go.AddComponent<CapsuleColliderCast>();
             go.AddComponent<ClientNetworkTransform>();
+
+            // Setup the rigidbody
+            Rigidbody rb = go.GetComponent<Rigidbody>();
+            rb.useGravity = false;
+            rb.isKinematic = true;
 
             // Setup animation controller.
             Animator anim = go.AddComponent<Animator>();
@@ -255,9 +262,6 @@ namespace nickmaltbie.openkcc.Tests.netcode.Runtime.Character
             capsuleCollider.center = new Vector3(0, 1, 0);
             capsuleCollider.height = 2.0f;
             capsuleCollider.radius = 0.5f;
-
-            // Get the components to update and modify.
-            capsuleCollider.enabled = false;
         }
     }
 }