+++ Packages/com.nickmaltbie.openkcc/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## In Progress
 
+* Updated kcc state machine to be synced with the network kcc.
 * Removed ParentConstraint requirement from the KCCStateMachine
     and replaced with a new `RelativeParentConfig` object to managed
     player's relative position to a parent object.